Consolidation into a single app
For ease of packaging, everything has been put into a single accessible
app. Now to package for installation
