+++ byte_orbit/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for weatherapp project.
+Django settings for byte_orbit project.
 
 Generated by 'django-admin startproject' using Django 1.11.3.
 
@@ -9,10 +9,6 @@ https://docs.djangoproject.com/en/1.11/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/1.11/ref/settings/
 """
-
-# superuser: roansong
-# email:     roansong@gmail.com
-# psswd:     byte_orbit
 
 import os
 
@@ -24,7 +20,7 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 # See https://docs.djangoproject.com/en/1.11/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'e)345ht^pe#9q6vxa9mqsep#p5ent%eo2t$bi7n^=ok5l0rqc&'
+SECRET_KEY = '7b-2#_j7g@vq6f!l@1$4wuoem9lge(svoindt_xh%+^0cmsm!+'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -43,7 +39,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'main.apps.MainConfig',
     'rest_framework',
-    'api',
 ]
 
 MIDDLEWARE = [
@@ -56,7 +51,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'weatherapp.urls'
+ROOT_URLCONF = 'byte_orbit.urls'
 
 TEMPLATES = [
     {
@@ -74,7 +69,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'weatherapp.wsgi.application'
+WSGI_APPLICATION = 'byte_orbit.wsgi.application'
 
 
 # Database
@@ -124,7 +119,7 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/1.11/howto/static-files/
 
-STATIC_ROOT = 'C:/Users/Roan\ Song/Desktop/jobhunt/projects/byte_orbit/weatherapp/static'
+STATIC_URL = '/static/'
 
 STATIC_URL = '/static/'
 LOGIN_URL = '/login/'
