+++ byte_orbit/urls.py
@@ -1,4 +1,4 @@
-"""weatherapp URL Configuration
+"""byte_orbit URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.11/topics/http/urls/
@@ -21,7 +21,6 @@ from main import views as core_views
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^',include('main.urls')),
-    url(r'^',include('api.urls')),
     url(r'^login/$',auth_views.login,{'template_name': 'main/login.html'},name='login'),
     url(r'^logout/$',auth_views.logout,{'next_page': '/'},name='logout'),
     url(r'^register/$',core_views.register,name='register'),
