+++ byte_orbit/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for weatherapp project.
+WSGI config for byte_orbit project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "weatherapp.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "byte_orbit.settings")
 
 application = get_wsgi_application()
