+++ main/templates/main/index.html
@@ -2,6 +2,6 @@
 {% block content_header %}Welcome!{% endblock %}
 {% block content %}
 <div id="content">
-	<a class="nice" href="{% url 'main:results' %}">View 7-day Forecast</a> <strong> OR </strong> <a class="nice" href="{% url 'api:create' %}">View API</a>
+	<a class="nice" href="{% url 'main:results' %}">View 7-day Forecast</a> <strong> OR </strong> <a class="nice" href="{% url 'main:create' %}">View API</a>
 </div>
 {% endblock %}
